requirements.txt
fastapi
uvicorn
yt-dlp
pystray
Pillow
psutil
packaging
requests
python-multipart